+++ BE/src/main/java/louie/hanse/issuetracker/web/dto/IssueSaveRequest.java
@@ -1,4 +1,4 @@
-package louie.hanse.issuetracker.web.dto;
+package louie.hanse.issuetracker.web.dto.issue;
 
 import java.util.ArrayList;
 import lombok.Getter;
